Close #1
Try yet another strategy for packing arguments in Msgpack RPC.
All arguments are of the form [{params}], and therefore
are listed by name.  This is true for 0, 1, and 2+
length argument lists.  The same consistent format
is used in all.

The issue with the array-based approach is that it's pretty
fragile if arguments are added or deprecated.

cc: @gabriel