Support for reading in devices when reconstructing sigchains
Keep track of Devices and Kid<->DeviceID mapping in ComputedKeyFamily.
Only do so for Sibkeys now.  That might be the wrong idea log-term,
but there are way to work around it.  In particular, every sibkey
should have a pointer to its recent subkeys, and we can then map
a DeviceSibkey to a DeviceSubkey, via computed key family