Close #313
Fixing #313 exposed an additional bug, which was that the eldest link
couldn't also delegate, but that was required in the case of a PGP key
delegating to a new device keyset.

There's still an additional bug, which is that the detkey is still
being provisioned by PGP, which may or may not be the right call,
but probably it's better to use the new device key for that.