add the AllKeys map to KeyFamily, and use that for FindKey
The Sibkeys and Subkeys maps that we get from the server do not contain
revoked keys. When validate a user's sigchain, we need to replay all the
signatures in order, and that means using keys that may now be invalid.
So we were running into errors when the user actually revoked keys,
because we couldn't find them to check old signatures.

Now we have the AllKeys map, which is built from data that was already
in the server response but not yet used. That includes keys that have
been revoked.