clean up KeyFamily
ServerKeyRecord contained a lot of untrusted fields from the server. All
that's really needed is Bundle, and everything else should be computable
from there, including KID and Fingerprint. Most of the rest has been
deleted, though we still get EldestKid for now, because getting rid of
it requires a server-side change.

We delete the sibkey and subkey lists entirely -- callers now have to
use ComputedKeyInfo to get a key's status, like they should.

Note that the KIDs in the AllKeys map are received from the server, but
we (always did) check that those match what we compute in Import.