libkbfs: clear MD block changes when loading MD from disk
This is just temporary until I implement proper copy-on-write for MD
in an upcoming commit.  Without this, the first write after a kbfs
restart could cause a crash, if the previous write before the restart
only ref'd/unref'd blocks in the top directory of a folder (because
the Children field would be an empty map, and so wouldn't be
serialized, and would be nil after deserialization).

Issue: #43