parse the eldest key from the merkle leaf
Previously we'd figure out the eldest key from the sigchain, but that's
ambiguous when the user has just reset their account. Now it's in the
Merkle leaf, so that it's unambiguous and hard for the server to lie
about it. This diff just parses the new leaf format -- actually using
the new field comes later.