kbfs: notification interface
The application layer on top of libkbfs (e.g., kbfsfuse), must have a
wayto get updates for changes that happen elsewhere (either in other
application layers on the same client, or from other clients
entirely).  This commit adds and implements Notifier and Notifiee
interfaces to make that work.

We now use this interface exclusively in kbfsfuse to be notified of
path updates, instead of relying on the updated paths returned from
KBFSOps.  In the future, we might want to not even return these paths
from KBFSOps (but currently the tests rely on them).

Issue: #21