kbfs: make RWChannel an interface, and add a simple inline impl
The KBFSOps tests run a significant amount of functionality inside
RWChannels, which use mocks.  Unfortunately, when mocks fail (due to
an unexpected call, for example), they call testing.FailNow(), which
is only supposed to be called from the main testing goroutine.  Thus,
any mock failures are just swallowed up, and end up causing deadlock.

Instead, these tests should use a simple version of RWChannel that
doesn't launch any goroutines, but just executes every request inline
in the main testing goroutine.

We should add tests in the future that check for concurrent access to
libkbfs, and use the regular RWChannel implementation for that.