fuse_ops: correct lookup errors in private, "fake" nodes
Looking up the public directory in another user's directory requires
creating a "fake" node for that user, with a nil directory ID.
FuseOps was trying to use this fake node to do real MD fetches, which
doesn't work since the ID is fake, but it returned a sketchy "leveldb:
not found" error.

Instead, FuseOps should check for the fake ID and return an
appropriate ReadAccessError.

Add a test to catch any regressions.