Move DirHandle parsing into libkbfs
This is useful for both FUSE implementations, and DirHandle.ToString
already understands the format, so this is logically the right place.

Simplify the internal logic while we're at it. Move UIDList into a
non-exported helper. Start introducing context where it makes sense,
even if there is no comprehensive support for cancellation yet.