kbfs_ops: properly copy dir and file blocks for writes
Without this, the archived version of the block gets updated.  If a
directory block might be re-accessed during an option (for example, if
syncBlockInChannel is called twice, as in Rename()), then it must be
stored back into the cache as a dirty block.  Try to minimize this by
moving most of the modification logic for mtime/ctime into
syncBlockInChannel.