fuse_ops: fix bug where MD updates weren't immediately reflected
FuseOps caches the directory entry, and updates it on a BatchChange.
However, it was only updating it when the node itself changes, whereas
the directory entry actually lives in the parent.  So when the
metadata changed for a file without the actual block changing for that
file (like when an mtime is updated), it wouldn't be reflected.

Also, add some comments on mtime/ctime.