Split KeyManager.GetTLFCryptKey() into three functions
This avoids the need to build a Path object just to pass into it.
It also makes sure that we always use the right key generation
(i.e., the latest one) for encryption.

Also have tests in kbfs_ops_test.go Verify that the right
RootMetadata object is passed into GetTLFCryptKey*.

This is part of #102. This closes #118.