libkbfs: introduce FolderBranchOps and new concurrency model
This commit reorganizes the KBFSOps code to execute within a
per-folder, per-branch instance.  This will make it easier to track
stuff on a per-folder, per-branch basis (such as the HEAD metadata,
and later the current local state of the folder-branch).

We also create a new concurrency model, using locks instead of
RWChannels.  This is because, to get good performance, we will need to
allow reads and file writes during remote-sync operations, and
previously syncs would block everything in our RWChannel model.
Instead, we use critical sections around state, as described in
design/state_machine.md.

Issue: #21