Plumb PassphraseGeneration() through to device update
Whenever we verify a passphrase against the server, we get back a generation
number.  It turn, we should present that back to the server whenever we're making
a change to LKS, since we desparately want to avoid race conditions in which we're
using the wrong (an old) passphrase for LKS.

Progress on #509.  All tests pass, with 2 massive caveats:
  1. We're not testing when generation #s are bumped up yet
  1. The tests just got really really slow (108s v 90s on my machine).
     I'm not sure what his happened