folder_branch_ops: allow writes and truncates during syncs
We want to allow writes and truncates to a file that's currently being
sync'd, like any good networked file system.  The tricky part is
making sure the changes can both: a) be read while the sync is
happening, and b) be applied to the new file path after the sync is
done.

This commit just does the dumb, brute force thing for now: if a block
is currently being sync'd, it copies the block and puts it back into
the cache as modified.  Then, when the sync finishes, it throws away
the modified blocks and re-applies the change to the new file path
(which might have a completely different set of blocks, so we can't
just reuse the blocks that were modified during the sync.)