folder_branch_ops: ensure that write/truncate always dirties top block
This commit ensures that any write or truncate operation dirties the
top block in the file.  Since the top block is always sync'd, this
ensures that every write/truncate call that comes in during a sync
gets deferred.

Also, there's no need to dirty the directory block on a write/truncate
anymore, since the deCache takes care of that.