kbfs_ops: take a Context in every KBFSOps method
This commit sets the groundwork for allowing cancelable commands, by
taking in a Context with every KBFSOps method.  Applications can then
cancel those contexts if they no longer want to wait for a method to
return.  (In a future commit, KBFSOps will actually listen for
cancels, but does not do so yet.)

This also allows applications to figure out whether a notification to
its Observer interface was caused by itself, via values in the
context.  In the future, we can also use this to tag and trace
individual operations as they move through the client and across RPCs.

Issue: #21