stop enforcing server ctime
We've discovered that the server sometimes has bad values for the ctimes
of old links. Rather than guarantee that the server always has this
right, we've decided to stop using the server ctime entirely, and just
get the signing ctime straight from the unverified sig payload. Since
this is what the Go client always did, the only change needed is to stop
checking the value against what the server says. (Which was done to make
sure the shared test suite passed against the Go code.)