libkbfs: call PutUnmerged on MD conflict, and track state
If the client tries to put an MD that has an out-of-date PrevRoot, the
MD server should reject it and the client needs to PutUnmerged() it.
The client also needs to track its state on a per-folder-branch basis,
in order to properly implement conflict resolution (coming in the
future).

Currently there is no way for a device to get out of the 'staged'
state once it's in it.  That will come later, after we implement
conflict resolution.

Issue: #21