libkbfs: make the notifier interface per-folder-branch
The caller should be able to specify on what branch to get updates for
a particular top-level folder.  For the foreseeable future, only the
'master' branch will be writeable, so this probably won't be used.
Still, it's cleaner to do this more explicitly.

At this point it makes sense to have a FolderBranch struct and expose
it externally for this interface, so this commit does that as well.

Issue: #21