FUSE: Keep track of active child nodes
This failed on OS X but not on Linux, because Linux rename seems to
add knowledge of the new path->nodeID mapping in the kernel as part of
the Rename operation, where as OSXFUSE performs a new Lookup for the
new path.

Previously, on OS X we'd see

- Lookup "old" returns NodeID 3
- Rename "old" to "new"
- Lookup "new" returns NodeID 5

Since there were two instances of File while only one logical file
existed, they ended up seeing different states of the file content,
and that's why the test failed. With this change, the second lookup
returns NodeID 3.

Avoiding this aliasing was a known missing functionality, I was just
waiting to have tests actually exercise it. Now, each Dir keeps track
of those children that the kernel holds references to.

Add a unit test for cross-directory renames while a file is open.

Fixes https://github.com/keybase/kbfs/issues/148