FUSE: Correct semantics for WriteRange truncation
The included unit test doesn't really trigger the edge case -- comment
out the `if write.Len==0` check to see -- as invalidating the
attributes of the file means Linux wouldn't try to read past the new
file length anyway. Perhaps it would be ok to not even invalidate the
lost tail of data; for now, we'll keep both the suspenders and the
belt, just in case.