Cleaning up track.
Created a TrackOptions record. Changed arg names to reflect ability to track with user assertions.

IdentifyUI finishAndPrompt would could change the options which was a little confusing. Simplified the code a bit (changed to “confirm”) and throwing an error if you try to track locally when you are tracking remote. (They would need to untrack first.)