FUSE: File Setattr returns attributes
This avoids an immediate Getattr callback, and makes things slightly
more efficient.

We need to explicitly set the Valid time, as bazil.org/fuse only puts
in the defaults on the Getattr side, or if Setattr is not implemented.
The value used here (1 minute) is what the default was already.
(FUSE API semantics may change to make Setattr easier to implement,
but us being more explicit here should never hurt.)