fix TestPGPUpdate
Fixes https://github.com/keybase/client/issues/532.

How we handle PGP keys has changed, and this test started failing. We
didn't notice because it was flaky before. Now it checks just the number
of PGP key bundles reported by the server (instead of looking for a
removed subkey -- because subkeys are now merged).