FUSE: Give EACCES not EIO on a non-existent folder we can't create
As far as file system semantics are concerned, this might make more
sense as ENOENT, but the error we see is very generic
WriteAccessError, and giving EACCESS in this scenario doesn't feel
horribly wrong. It's not that it doesn't exist, it's that you can't
automatically create it by accessing it. All valid folder names always
"exist" and are instantiated on demand.