bcache: add a way to lookup BlockIDs based on hash of plaintext
In order to deduplicate blocks, the client needs to be able to figure
out if a particular block is already stored on the server.  However,
we don't want to give the server the plaintext hash of the contents,
and the encrypted hash (i.e., the BlockID) is random, even with the
same contents.

So we implement a poor man's way: keep an LRU cache of hash -> ID
mappings on the client side.  So in the common case of the client
duplicating a file it just read, we can be pretty confident that the
blocks are still in the local cache and we can find the ID again.

Note that this is only useful (and currently restricted to) direct
FileBlocks, because anything containing BlockPointers to other nodes
will have to deal with changing RefNonces and won't ever have the same
plaintext hash.  We also restrict deduping to within the same folder.

Issue: #37