libkbfs: add codec extension for iface slices, and use for ops
The ugorji codec doesn't support decoding data into interface types;
it requires either a hard type or interface{}.  But we really want
BlockChanges.Ops to be typed appropriately.  So, this commit adds a
codec extension that does type conversion from the interface{} type to
the actual type expected by the slice in question.

Unfortunately, the ugorji codec also doesn't allow you to decode
extensions into pointer types.  Sigh.  So all the extension types end
up being non-pointers on a decode, which doesn't satisfy the
interface!  Fix this in a somewhat hacky way by passing in a function
to do the type assertion.  (Ideally, we'd be able to do generic type
assertions via the reflect library, but I don't think that's
possible).