Cache the armored version of PGP keys
Go's randomized iteration order means that every time we serialize an
openpgp.Entity we can get different results (headers and packets in a
different order). Caching the armored version is a little dangerous
since we'd have to invalidate it if we ever modify the entity, but it
works for now.