port Chrome RPM postinst file
Like Debian, RPM Chrome installs a cronjob by default. I'd like to avoid
that, so I've picked around it. Unlike Debian, it seems that cronjob
might be mandatory for non-Yum-based RPM distros (OpenSUSE, Mandriva).
I'll test it to see if that's still true, and we can decide whether we
care enough to port over the extra complexity.