Progress on #680
This bug is likely limited to standalone mode, OR!!! in server mode
when the user doesn't have any data in their passphrase stream cacche.

The problem is that if the passphrase stream cache is empty, we just
bail out entirely on attempting to load secret keys. Clearly this
is a bug, but I'm not sure what was intended. Need input from @patrickxb

Note I think that even if we fix this, the behavior of #681 is still
broken