Make TestKBFSOpsConcurWriteParallelBlocksCanceled more deterministic
A worker that processes the initial two blocks may or
may not exit before blocking on the next one, leading
to a race condition with the cleanup loop. This gets
rid of that by making all workers block on the ready
channel before cancelling.

This is part of #281, and a follow up to #288.

This closes #281. Thos closes #292.