forward log messages from the daemon to clients
The goal is that all calls to G.Log.* should work whether you're in
client or daemon mode. In an ideal world, every function that wanted to
log something would have access to a context object with
connection-specific info. In the current world, we have to make some
global hacks to make this work, because many functions that want to log
do not have a context. Lots of moving parts here:

- The Logger.Standard implementation now stores a list of external log
  implementations. Any time it logs something, it also forwards logs to
  these external loggers.
- Whenever a client connects to the daemon, the daemon instantiates a
  LogUI for the life of that connection and registers it as an external
  logger.
- To avoid sending a lot of unnecessary data over the IPC pipe, the main
  logger stores a log level, below which the external logging will be
  skipped.
- When starting up in client/daemon mode, the client now immediately
  makes a SetLogLevel call to set the level on the daemon side
  (generally to INFO or DEBUG). This runs the risk of multiple
  simultaneous clients stomping on each other's log mode (as well as
  seeing each other's log events), but since this is hopefully a
  temporary hack we're not going to worry about it too much.