folder_branch_ops: use new logger
Define a module that prints the TLFId with every log message, and use
it print the start/exit of every major method, including the nodes
involved and any resulting error.

When FOlderBranchOps receives asynchronous events, like MD update
notifications, spawn a new request ID as a tag for all the resulting
work.

Issue: #5