libfuse: don't lock during observer notifications
Locking during notifications can lead to deadlocks, especially for
oxxfuse 3.X which makes a synchronous GetAttr call during
invalidations.

Instead, create a background goroutine that delivers notifications
asynchronously (but in order).  This is effectively an unbounded
queue, but that can't be avoided.

Also add a Waitgroup for notifications (suggested by @tv42) for tests
to know that delivered notifications have all been acted on.

Issue: #331