Narrow the scope of libfuse.Folder.mu: Easy cases
Now that libkbfs is node-based, we have stable references to the
underlying nodes, and don't have to worry about managing pathname
mappings. This removes most of the need for the Folder mutex. It
will remain to guard the `fs.Node` to `libkbfs.Node` mapping,
only.

Rename attrLocked as there's no longer anything really kept
locked.