Narrow the scope of libfuse.Folder.mu: Push lock into forgetNode
Now that libkbfs is node-based, we have stable references to the
underlying nodes, and don't have to worry about managing pathname
mappings. This removes most of the need for the Folder mutex. It
will remain to guard the `fs.Node` to `libkbfs.Node` mapping,
only.

As the callers no longer have any need for the locked, push it inside
the method.