In TestPGPUpdateMultiKey, actually do the updates
Previously, this test just checked that the arguments were interpreted
correctly. Since the generated keys were never exported to the local
keyring, the update did nothing.

Since the process of updating PGP keys is a little more complicated now,
and a little more security-conscious, it's useful to test the whole
thing.