Check passphrase stream exists before device prov
The passphrase stream cache needs to exist before a
device is provisioned.  Locksmith had an error for
this situation before, but it left the user in an odd
state.

With these changes, it ensures that the passphrase
stream is available before attempting to device
provision or repair the user's device keys.

Also, the login engine detects a missing passphrase
stream before trying to use Locksmith to provision a
device.  If it's missing, it logs the user out and tells
them to log in again.

This should fix #872.

One way to get into this state is:

1. The user is logging in for the first time on a device.
2. Before the device is provisioned, the login is canceled or
   interrupted, but they do successfully get a login
   session.
3. The daemon restarts (`ctl stop`, machine reboot, bug, etc.)
4. The login session is still valid, so the next login attempt
   does not require a passphrase, and the passphrase
   stream has been lost due to daemon restart.

There's a new test for this scenario: TestLoginRestartProvision
It exhibited similar behavior to the stack trace in #872
and is fixed by these changes.