Abort pubkey login if no synced devices exist
Fixes #926

Before, after `keybase reset`, `keybase login` was
trying pubkey login.  Pubkey login was unlocking lksec
to find a device key when there were no keys synced.

Thus it was asking for a passphrase and failing to find
a device.  Since there was a failure, the passphrase
stream was not saved since failures signify verification
problems.

So then after pubkey login failed, it asked for a
passphrase again.

This commit aborts pubkey login if there are no synced
devices.