ops: add the BlockPointer being renamed to renameOp
For conflict resolution, we need to be able to find the complete paths
to a given BlockPointer, using only a starting MD revision and a
subsequent set of operations.  Without knowing where a BlockPointer
moved, there was no clean way to do that for rename operations.  So
this commit just saves the BlockPointer in the renameOp.

Issue: #321