On load from storage, don't store if no results
Fixes the synced pgp download part of
keybase_issues#1783.

If a load from local storage returned nothing, it was
setting the synced keys to an empty set with version 0.
This caused the subsequent api request for version 0 of
the user's private keys to fail.

Old users don't have a version set in the versions
table.  So a fetch_private call with version 0 failed
to return anything.

Now, the first sync won't include the version parameter,
so it will work.