kex2: duplex secure channel
This commit introduces a kex2 secure channel.  The function of the server is mocked-out
via the MessageRouter interface.  Eventually we can supply API endpoints on the server
to handle the implementation here.

Tested pretty well but there could be bugs.

The next step is to build an RPC layer on top of this, and then circle back to find
any bugs that crop up along the way.

Closes #1015

From review:

Address @patrickxb's comments, and play it safe w/r/t hmac Equal