move generated Go protocol code from go/vendor/... to go/protocol/
The protocol Makefile has been repointed to the new directory. This
means we no longer have any of our own protocol code under
client/go/vendor/.

Because of the path change, every file importing the protocol
definitions had to be updated. I did this with sed:

    sed -i 's%"github.com/keybase/client/protocol/go"%"github.com/keybase/client/go/protocol"%' **/*.go

This will break OSX and other code that depends on this repo too.