folder_branch_ops: save MD during CheckForNewMDAndInit
The first time a folder is accessed, KBFSOpsStandard directly fetches
the MD object for a TLF from MDOps (because it doesn't know the TlfID
yet).  It calls FolderBranchOps.CheckForNewMDAndInit() to initialize
the MD if necessary.  If the MD doesn't need initializing,
FolderBranchOps would just return -- after kicking off the
notification process, but without saving the MD anywhere.  As a
result, the notification process would just start downloading all
revisions, because it doesn't know what the current one should be.

Instead, save the MD before doing anything else.  This requires
passing in the unmerged MD, if one exists, so try fetching that first
in KBFSOps.  Also, to avoid code duplication, consolidate the
saveMdToCacheLocked method into setHeadLocked().

Issue: #404