folder_branch_ops: don't delete syncOp until after dirty deletes
If some blocks fail to put during a Sync(), we would leave dirty
blocks in the cache, but the corresponding syncOp would have already
been cleared by mergeUnrefCacheLocked.  If Sync is then called again
on the same file without any intervening writes, there would be a
crash because the code assumes a syncOp will be there if there are
dirty blocks.

But we don't need to delete the syncOp there, since we already delete
it later after the blocks are successfully put.  So just get rid of
this delete entirely.

Issue: #422