folder_branch_ops: don't take cache lock during unstaging
Inverting some local operations (such as setattr) require looking up
directory entries, which requires potentially replaying writes from
the cache, which requires taking the cache lock.  So this results in a
deadlock if UnstageForTesting also takes the cache lock.

I believe it's safe to remove to let writes in during the unstaging;
they haven't been committed yet so they will just be applied to the
next version of the file.

Issue: KBFS-464