Make a streaming base58-encoder/decoder
Mainly take the template from Go's standard base64 encoding library,
but add some tweaks for our setting.

Introduce the idea of 'Strict' mode in which all characters must be
from the base58 alphabet, and implicity, liberal mode is when all non-alphabet
characters are implicitly ignored.

Add in a filteringReader to handle the streaming of bad characters

Use the Go library tests to test it all, but we should probably write some
more tests too