stop trying to create a prompter in `keybase signup --batch`
Because this is done as part of ParseArgv(), it happens before the
global InitUI() call is made in mainInner(). That means that UI is nil
and GetTerminalUI() throws.

As far as I can tell though, we don't need this line at all. Setting
doPrompt=false prevents us from ever using it.