A CredentialAuthority for mapping UID<->Username<->KIDs
We should be able to query this thing repeatedly. It will manage caching
and make sure that we don't slam the server. It will listen to pubsub pushes
to eagerly timeout users, and otherwise will just time them out after a reasonable
delay (~5 minutes as currently configured).

Tests for a mocked-out server interface. The real API compiles but has yet
to be tested.