move ClearStoredSecret back into the DeprovisionEngine
When I transmorgrified ResetEngine into DeprovisionEngine, I moved the
ClearStoredSecret call into G.Logout(), on the assumption that that was
something we should always be doing when we log out. That broke a bunch
of tests, and now that I have a better understanding of how we cache
secret keys in memory, I'm not sure it was the right move at all. I'm
putting it back.