Fix unknown username bug in login via passphrase
In particular, for web users without keys logging in.

https://github.com/keybase/keybase-issues/issues/1855

This also fixes another bug where logging in on a device
with another account would try to get the passphrase for
the original account instead of for the username
specified during login (either via UI or command line).

Added/clarified tests for both situations.  Verified
that they fail without fix, pass with fix.