delete a bad check for key ownership
https://github.com/keybase/keybase-issues/issues/1878

Some old sigchain verifying code was incorrectly assuming that either
1) You have a self-signing PGP key, or
2) You have an eldest link of an explicit eldest type.
But it's possible for accounts created using an older version of the
website to be in a state where neither is true.

In fact, we don't need to check key ownership at all in this part of the
code. The relevant check is done by InsertServerEldestKey on a different
code path. We also have a test vector that covers this case
('test_error_empty_sigchain_bad_key_userid'). So we can cleanse the
earth with fire here.