Ignore lksec error during login if there's no device ID
During LoginWithPrompt, it's possible for the user to
select secret storage and it tries to store the secret
right after establishing a login session with the API
server.

This works fine on a provisioned device, but if the
LoginWithPrompt is being called to establish a login
session for device provsioning, then there is no device
ID yet, and thus secret storage is not possible.

Before this change, LoginWithPrompt would return an
error and provisioning would bail out.

As other secret storage errors are ignored, it seems ok
to ignore this error as well.  The most important thing
is to get the device provisioned.

It would be nice to turn off the checkbox for secret
store in this case.  That would require more code
changes.  This is a minimal change to fix the bug.

Fixes part of https://github.com/keybase/keybase-issues/issues/1822
where there is a "no device id set" error.