cr: proper block accounting for reinstated files
When a file is deleted in the merged branch, and then reinstated by
the unmerged branch, we should make sure that the corresponding
createOp adds ref blocks for all the newly reinstated blocks,
especially for multi-block files.

Moreover, if a file is being reinstated because of a rename, the
remote notification should be a createOp, not a renameOp.

Issue: KBFS-517