spec changes to stop misusing Poly1305
Previously we were authenticating for each recipient the Poly1305 tag of
each payload secretbox. This was a mistake. A polynomial authenticator
isn't a hash, and other recipients who know the session key will be able
to find other ciphertexts that authenticate with the same tag.

Instead, authenticate for each recipient the SHA512 of the whole payload
secretbox. This should provide the authenticity we thought we had
before.