folder_branch_ops: defer all writes to a block during a sync
Previously the first write to a block would clear the block out of
`copyFileBlocks`, which means that subsequent writes happening during
the same sync would not be deferred (and would be lost after the
sync).  This is because we were conflating two things: the need to
copy the block, and the need to defer writes.

Instead, track multiple states for blocks being sync'd so that we can
tell the difference between the two.

Issue: KBFS-616