clear the cached secret key state before running SecretKeysEngine
This is a minor safety measure, so that we don't expose cached secret
keys to all callers without the user's consent. We're not generally
trying to be secure against malicious code running locally (on the
assumption that at that point you're already screwed), nor are we
generally willing to rely on the user's password for security (on the
assumption that it's short and easy to guess), but it doesn't cost us
much to be conservative in this case.