Fix retrack
The Retrack button calls trackUser() immediately, which checks
`trackerState && trackToken && shouldFollow` when deciding whether
to perform a track.

`shouldFollow` is set by the Track checkbox that we show under
trackerState == normal.  So in the Retrack/Untrack case, it's not
true because it was never shown, and we do nothing when you try to
Retrack.

This PR fixes it by only testing shouldFollow if we're showing a
"normal" trackerState.  If we're showing a warning/error state,
we don't need to test shouldFollow.