folder_branch_ops: reinstate full block changes in MD after Put()
It seems best to keep the full set of block changes around locally in
the MD object.  We do this whenever we download an MD from the server
and re-embed the block changes, so we might as well also do this when
the MD originates locally.

This will simplify block archiving logic by letting us pass around
just the MD object to the background archiver, without worrying that
the cached block changes will get wiped out from under us.

Issue: KBFS-485