folder_branch_ops: reset syncOp state if Sync() call fails
If a Sync() errors out after the block changes are filled in on the
given syncOp, but before the MD is officially Put() (due to a block
Put error, for example), the syncOp will be reused for the next Sync()
call.  However, it will still contain all the block changes from the
previous failed attempt, which will be wrong!  Thus, after the next
Sync(), the syncOp could have two updates for each BlockPointer, which
ends up destroying the block history and can make the directory
entries and the node caches out of sync.

So, reset the block change state of the syncOp if there's ever an
error.

Issue: KBFS-635