Fix KBFS not spawning pinentry
We were seeing something strange:

  start service -> start electron -> start kbfsfuse

caused a pinentry popup asking for your Keybase login for KBFS,
which is good, but:

  start service -> start electron -> restart service -> start kbfsfuse

did not cause any popup, and KBFS sat failing silently.  It turns out
that the difference was the config.getCurrentStatus() RPC that we
called only once at Electron startup time.

We will fix the core bug in a separate PR too, but Patrick says that
having us reissue the getCurrentStatus RPC when we reconnect to the
service is orthogonally a good idea.