Fix Linux bringing up the splash screen at every boot
On Linux, we want to store the started.txt file in the data dir
because the socket dir is ephemeral.

On Windows, we want to store the started.txt in the data dir
because the socket dir isn't actually a filesystem.

On OS X, we want to store the started.txt in the socket dir,
which launchd manages for it.