Forbid deleting the alias symlinks in top-level folder lists
Previously, trying to remove `private/jdoe,jdoe` (or any such name
where the file was a symlink to a canonical top-level folder) caused
an `Input/output error` (EIO). Give a nicer error. EPERM is the error
number to use when the operation is impossible in a filesystem.

The user-facing error now looks like:

    rm: cannot remove ‘.../private/foo,bar’: Operation not permitted

Issue: KBFS-469
Signed-off-by: Tommi Virtanen <tv@eagain.net>