Support hidden remote components.
Hide tracker while doing the tracking rpc request.

Currently we closed the tracker after starting the tracking request. But
that lead to errors like: https://keybase.atlassian.net/browse/DESKTOP-277
and possibly a race condition(?).

This will hide the window while we're making the request, and close it
when it's done or errors.

We can refine this later to have the same remote component come back up
with a helpful error message.