Don't reset the close state when we get another identify call.
The service can make multiply identify calls to the app for the same
user. For example when the user goes to `private/user,foo` and `private/user,foo,bar`
That will yield two identify calls to `foo`. So instead of closing the
tracker popup and reopening it when we've decided to show it again,
let's just leave it open.

This will mean that the tracker popup will update to say "Checking" if
we get a launchNetworkCheck or finishProof from the service.

Ideally the service would only send updates, and not reset the state
with launchNetworkChecks.