Change external logging to use queue per connection
1. Moved external logging code out of logger package.
2. service has a single log forwarder that receives log
messages from the logger package.
3. New log protocol for RegisterLogger.  Clients have to
call this if they want service log messages.
4. Have CLI call RegisterLogger.
5. Each connection creates a log queue.  If
RegisterLogger is called, a channel is created with a
processing goroutine.

Bonus: in debug mode, added a small timeout before the
client shuts down to allow it to receive all the log
messages.  Also, the log queues cleanly shut down so all
the messages are sent.