folder_branch_ops: use exponential backoff for MD updates
If we fail to process an MD update, retry with exponential backoff
rather than immediately.  If, for example, the MD given to us by the
server is bogus, we don't want to keep pegging the MD server (or our
own CPU) repeatedly hitting the same error over and over again.

Issue: KBFS-666