folder_branch_ops: fix pointers in file top-blocks after error
When we hit a recoverable block error on a multi-block file, we must
make sure that any outstanding dirty versions of the file are fixed up
to reflect the fact that some of the indirect pointers now need to
change.  Without this, it's possible for a deferred write to make a
copy of the top block that contains a pointer that has been archived,
and when that deferred write syncs, no non-writer will be able to read
the file.

Issue: KBFS-700