conflict_resolver: only recreate operations on deleted dirs
When a directory is deleted on an unmerged branch, but modified on the
merged branch, the directory will survive, but we still want to replay
all the entry deletes that happened while unmerged.  But for a file,
replaying syncs and setattrs doesn't make any sense and causes issues.

Issue: KBFS-668