Fix path displayed in tracker popups
For public directories, it would show
"/keybase/public/foo#public". Add
GetCanonical{Name,Path}() methods
to TlfHandle and use that instead.

Also fix the rekey notification to
use the canonical path, too.

Define a CanonicalTlfName type so
that we can keep track of where
we're using canonical names.

This closes #665.