use libkb constants for our log file names
Previously these were defined in several places, but it's important to
have them stay consistent. As part of deduplicating these, get rid of
the ability to configure them. We can add that back later if we decide
we need it, but using custom log paths is inevitably going to mess with
`keybase log send`, when someone runs that with a different config.