Update Java for android React Native
They've changed how to implement native components. Specifically
updateView is no longer a thing. I haven't looked into the changes yet,
so the custom TabBar is disabled for now.

We may be able to do without the custom tab bar (with the native view
pager) but making custom components will be pretty critical.