Fix "Close" on tracker logged out popup causing track to happen
We have a "Close" button appear on different types of tracker popup,
and then convoluted logic for working out which type of "close" this
is and whether it should result in a track.  This caused a bug where
the "logged out" tracker popup would track people when closed.

This PR fixes that by switching to a "tracker:onClose" (which closes
the popup without taking any action) -- which is now used both on the
header component and the action bar component in the logged out case --
and a "tracker:maybeTrack" action, which triggers the convoluted logic
I mentioned earlier.  We only call maybeTrack when tracking is a
possible outcome of the button click.

This flow is going to be thrown out and replaced by simpler callbacks
as part of the DZ2 rework, so we don't need to get too caught up in
this, but this seems like a simplifying fix until we get to DZ2.

Tested on stage0 using signup of real accounts, in the logged out
and logged in cases.  But please review carefully anyway, since it's
important to get right.