cr: fix resolution when there are no merged updates
When the only merged update is something like a rekey request, there
will be no changed block pointers, and thus the `mostRecent` ptr
returned from `resolveMergedPathTail` will be uninitialized.  In that
case, no resolution is needed, so skip it.

Add a test to catch this in the future.

Issue: KBFS-773