Speed up slow test
TestSecretboxEncryptedLen was taking ~1.5s,
which is a good chunk of the total libkbfs
test runtime (~3.7s).

It turns out it was because it was reading
a lot of random data needlessly. Change it
to not do so.

Also do the same thing for TestBlockEncryptedLen,
which was suffering from the same problem to
a lesser extent.

This closes #724.