libkbfs: loosely cache Favorites
Clients shouldn't call `KBPKI.FavoriteAdd` if a folder is already in
their favorites list.  To achieve that, `KBFSOps` now fetches and
caches the favorites list, and checks it before making any Add calls.

Manage this cache via a separate class and goroutine, so we can keep
it consistent in the face of adds and deletes.

Note that the cache is never invalidated by the server, so in theory
it's possible that a client will fail to re-add a favorite that has
since been deleted by another device.

Issue: KBFS-797