libfuse: don't set a timeout on the notification processor (OSX)
Because we used `WithContext` on the context we passed to
`fs.Serve()`, it had a deadline attached to it on OSX.  This meant
that our notification processor stopped after 19 seconds!  We don't
need the op ID or anything else set, so just set the app ID directly
rather than call `WithContext`.

Issue: KBFS-824