always make fresh clones in Docker, instead of copies
Previously we made copies so that it would be possible to do a Docker
build with local changes. Now, I find that I rarely do that, and the
copies cause confusion. (For example, the kbfs repo needed to have
master checked out or it would crash the build.) Instead, make fresh
clones, and use git's --reference flag to make the clones very
efficient.